modularized SignalChart and SignalTable into seperate components
